Extract ensembl module from the monolith (split PR 6/N)
Moves ensemblrest() and ensemblrest_rate() into
tfbs_footprinter3/ensembl.py.

The monolith continues to re-export both names so the HPC cache
monkey-patch at hpc/ensembl_cache.py::patch_tfbs_footprinter3() keeps
working: all current callers of ensemblrest() live inside the monolith
and resolve the name via its module globals, which is exactly what
the patch rewrites. Verified end-to-end by running a live Ensembl REST
call through the patched wrapper and confirming it returns
species='homo_sapiens' for ENST00000361390.

A note in ensembl.py documents the contract: if any future PR moves
an ensemblrest caller OUT of the monolith, the patch target needs
updating in lockstep (either switch callers to attribute-access via
, or
extend the patch to rebind both namespaces).

Monolith: 1271 -> ~1200 lines. 21 tests green, ruff clean.

Remaining in monolith (~1200 lines):
- experimentalDataUpdater, experimentaldata, species_specific_data
  (~300 lines) — data_loader concerns
- tfbs_finder, alignment_tools, retrieve_genome_aligned,
  transfabulator, test_transcript_id, transcript_data_retrieve,
  gene_data_retrieve, retrieve_regulatory (~300 lines) — pipeline
- main, get_args (~300 lines) — cli
- remainder: module-level state (script_dir, curdir) and re-export block
